Move warning message and validation of fields to widget object
